rename: HttpVersion 패키지 변경
