+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11.request;
 
+import org.apache.coyote.http11.HttpVersion;
 import org.assertj.core.api.*;
 import org.junit.jupiter.api.Test;
